Organise item and itemlist into packages to improve code quality
Fixes #257
