+++ src/main/java/seedu/duke/ListContainer.java
@@ -5,6 +5,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
 
+import seedu.duke.lists.ItemList;
 import seedu.duke.exceptions.HotelLiteManagerException;
 import seedu.duke.storage.AssignmentListFileManager;
 import seedu.duke.storage.RoomFileManager;
