+++ src/main/java/seedu/duke/command/itemcommands/DeleteItemCommand.java
@@ -2,11 +2,11 @@ package seedu.duke.command.itemcommands;
 
 import seedu.duke.command.Command;
 import seedu.duke.exceptions.HotelLiteManagerException;
-import seedu.duke.Item;
+import seedu.duke.lists.Item;
 import seedu.duke.exceptions.EmptyItemNameException;
 import seedu.duke.ListContainer;
 import seedu.duke.Ui;
-import seedu.duke.ItemList;
+import seedu.duke.lists.ItemList;
 import seedu.duke.storage.ItemListFileManager;
 
 import java.io.IOException;
