+++ src/main/java/seedu/duke/command/itemcommands/UpdateItemPaxCommand.java
@@ -4,8 +4,8 @@ import seedu.duke.command.Command;
 import seedu.duke.exceptions.HotelLiteManagerException;
 import seedu.duke.exceptions.EmptyItemNameException;
 import seedu.duke.exceptions.EmptyItemPaxException;
-import seedu.duke.Item;
-import seedu.duke.ItemList;
+import seedu.duke.lists.Item;
+import seedu.duke.lists.ItemList;
 import seedu.duke.Ui;
 import seedu.duke.ListContainer;
 import seedu.duke.exceptions.InvalidItemPaxException;
