+++ src/main/java/seedu/duke/command/itemcommands/ViewItemListCommand.java
@@ -3,7 +3,7 @@ package seedu.duke.command.itemcommands;
 import seedu.duke.command.Command;
 import seedu.duke.ListContainer;
 import seedu.duke.Ui;
-import seedu.duke.ItemList;
+import seedu.duke.lists.ItemList;
 import seedu.duke.exceptions.HotelLiteManagerException;
 import seedu.duke.exceptions.InvalidViewItemsCommandException;
